refactor: extract shared UI support from KgsCalendarApp
